e2e: one harness — fold credeval in, and add the model-behaviour suite
Every eval now runs through rustykrab-e2e: one binary, one report, one
exit code. `rustykrab-credeval` is removed; its scenarios, surfaces,
capture server, and outcome classifier move here intact.

Three modes, which the report deliberately keeps apart:

  scripted    replayed script, no model — boolean assertions, must pass
              every run, gates CI
  model       real model with stubbed tools — assertions plus an LLM
              judge, must pass a majority of repetitions
  credential  real model, real tools, empty credential store — an outcome
              distribution per surface, gates nothing

The third is why this is not a straight merge. e2e is a gate; credeval
was a measurement, with no process::exit anywhere, because its product is
a distribution: which of several distinct things the agent did when it
lacked a credential, at what rate, on which surface. You can derive a gate
from a distribution but not the reverse, and forcing these into pass/fail
would throw away the thing they exist to produce.

So Expected grows a third case, Measure. An xfail is boolean by
construction — it asks whether something passed when it should not have —
and a rate has no such thing. A threshold dressed up as an xfail would
make the report lie about what was measured. A measured scenario reports
its distribution and never turns the suite red; give it a threshold when
you want it to gate.

Also here:

- Backend::Model carries an optional channel, so a scenario can be driven
  over Telegram or Signal rather than only the gateway. The capture server
  stands in for the Telegram Bot API and signal-cli with no network
  egress, answering both long-poll shapes — reply with the wrong one and
  the daemon's poll loop logs a decode error every second for the length
  of the trial.
- The trial sidecar and --resume are preserved. A full credential run is
  hours long and the summary is written only at the end; losing all of it
  to a SIGTERM is not hypothetical.
- The classifier gains the tests it never had. Its priority order is load
  bearing — a filed request outranks the prose, fabrication outranks a
  polite ask — and nothing pinned that down.
- An empty tool-stub spec now means "leave the registry alone" rather than
  writing an empty file the daemon then fails to parse. The credential
  suite needs the real tools: its premise is that they cannot run without
  a credential.

The 17 scripted scenarios are unchanged and still pass.
