--- crates/rustykrab-credeval/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "rustykrab-credeval"
-description = "Behavioural eval: does the agent ask the user for credentials it does not have?"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[dependencies]
-tokio.workspace = true
-tokio-stream.workspace = true
-reqwest = { workspace = true, features = ["stream", "json"] }
-serde.workspace = true
-serde_json.workspace = true
-anyhow.workspace = true
-axum.workspace = true
-tempfile = "3"
-# Direct reads of the daemon's store, so a trial is scored on whether a
-# credential request row actually exists — not on what the agent claimed.
-rusqlite = { version = "0.34", features = ["bundled"] }
