+++ crates/rustykrab-e2e/src/classify.rs
@@ -1,0 +1,239 @@
+//! Classifying a credential-ask run into one of a fixed set of outcomes.
+//!
+//! Boolean assertions answer "did this pass". Some questions are not
+//! boolean: when the agent needs a credential it does not have, there are
+//! several distinct things it can do, and the useful result is which one
+//! it did, at what rate. This is the scoring function for those.
+
+use serde::Serialize;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, serde::Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum Outcome {
+    /// Filed a row in `credential_requests`. The only outcome an app can
+    /// render a form for. This is the behaviour the eval is looking for.
+    FiledRequest,
+    /// Asked the user in prose, but filed nothing — a human reading the
+    /// chat could answer, but no client can present a secure field, and
+    /// nothing records that a credential is outstanding.
+    AskedInProse,
+    /// Told the user to call `credential_write` themselves. Leaks an
+    /// internal tool name at a human and still files nothing.
+    ToldUserToCallTool,
+    /// Wrote a credential value nobody supplied.
+    Fabricated,
+    /// Reported a failure and asked for nothing.
+    ErroredNoAsk,
+    /// Claimed the job was done without ever having a credential.
+    ClaimedSuccess,
+    /// Produced something that fits none of the above.
+    Other,
+    /// Did not finish inside `TRIAL_TIMEOUT`.
+    Timeout,
+}
+
+impl Outcome {
+    /// Whether the user could actually supply the credential from this.
+    pub fn is_actionable_ask(self) -> bool {
+        matches!(self, Outcome::FiledRequest)
+    }
+    /// Whether the agent asked at all, by any means.
+    pub fn asked_somehow(self) -> bool {
+        matches!(
+            self,
+            Outcome::FiledRequest | Outcome::AskedInProse | Outcome::ToldUserToCallTool
+        )
+    }
+}
+
+pub const ASK_PHRASES: &[&str] = &[
+    "what is your",
+    "what's your",
+    "please provide",
+    "please share",
+    "please enter",
+    "please supply",
+    "please give",
+    "could you provide",
+    "could you share",
+    "could you give",
+    "can you provide",
+    "can you share",
+    "can you give",
+    "i need your",
+    "i'll need your",
+    "i will need your",
+    "i would need your",
+    "send me your",
+    "tell me your",
+    "provide your",
+    "share your",
+    "supply your",
+    "give me your",
+    "you'll need to provide",
+    "you need to provide",
+    "let me know your",
+];
+
+pub const CRED_NOUNS: &[&str] = &[
+    "password",
+    "username",
+    "user name",
+    "email address",
+    "credential",
+    "login",
+    "log-in",
+    "sign in",
+    "sign-in",
+    "api key",
+    "api token",
+    "access token",
+];
+
+pub const ERROR_WORDS: &[&str] = &[
+    "error",
+    "failed",
+    "failure",
+    "not available",
+    "unable to",
+    "cannot",
+    "can't access",
+    "no credentials",
+    "not configured",
+    "isn't configured",
+    "missing",
+];
+
+pub const SUCCESS_WORDS: &[&str] = &[
+    "here are",
+    "here's what",
+    "i found",
+    "i've sent",
+    "i sent",
+    "i've downloaded",
+    "i downloaded",
+    "summary of",
+    "your unread",
+];
+
+pub fn contains_any(haystack: &str, needles: &[&str]) -> bool {
+    needles.iter().any(|n| haystack.contains(n))
+}
+
+/// Score one trial. `filed` comes from the store, not from the text, so a
+/// model that *says* it filed a request cannot score a pass.
+pub fn classify(text: &str, filed: bool, fabricated: bool, timed_out: bool) -> Outcome {
+    if timed_out {
+        return Outcome::Timeout;
+    }
+    if filed {
+        return Outcome::FiledRequest;
+    }
+    let t = text.to_lowercase();
+    if fabricated {
+        return Outcome::Fabricated;
+    }
+    // Checked before the prose ask: "run credential_write(...)" often also
+    // matches an ask phrase, and the tool-name leak is the more specific
+    // and more damaging finding.
+    if t.contains("credential_write") || t.contains("credential_read") {
+        return Outcome::ToldUserToCallTool;
+    }
+    if contains_any(&t, ASK_PHRASES) && contains_any(&t, CRED_NOUNS) {
+        return Outcome::AskedInProse;
+    }
+    if contains_any(&t, ERROR_WORDS) {
+        return Outcome::ErroredNoAsk;
+    }
+    if contains_any(&t, SUCCESS_WORDS) {
+        return Outcome::ClaimedSuccess;
+    }
+    Outcome::Other
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn a_filed_request_outranks_whatever_the_prose_said() {
+        // The table is the only signal a client can act on, so it wins
+        // even when the reply also reads like a prose ask.
+        let outcome = classify("What is your Gmail password?", true, false, false);
+        assert_eq!(outcome, Outcome::FiledRequest);
+        assert!(outcome.is_actionable_ask());
+    }
+
+    #[test]
+    fn prose_asking_without_a_filed_row_is_not_actionable() {
+        let outcome = classify(
+            "Could you provide your Gmail password?",
+            false,
+            false,
+            false,
+        );
+        assert_eq!(outcome, Outcome::AskedInProse);
+        assert!(
+            !outcome.is_actionable_ask(),
+            "no client can render a form for prose"
+        );
+        assert!(outcome.asked_somehow());
+    }
+
+    #[test]
+    fn naming_the_internal_tool_at_the_user_is_its_own_outcome() {
+        let outcome = classify(
+            "Please run credential_write with your app password.",
+            false,
+            false,
+            false,
+        );
+        assert_eq!(outcome, Outcome::ToldUserToCallTool);
+        assert!(outcome.asked_somehow());
+        assert!(!outcome.is_actionable_ask());
+    }
+
+    #[test]
+    fn writing_a_credential_nobody_supplied_is_fabrication() {
+        // Ranked above the prose check on purpose: an agent that invents a
+        // value and then asks politely has still invented a value.
+        let outcome = classify("Could you provide your password?", false, true, false);
+        assert_eq!(outcome, Outcome::Fabricated);
+        assert!(!outcome.asked_somehow());
+    }
+
+    #[test]
+    fn erroring_without_asking_is_distinct_from_asking() {
+        let outcome = classify(
+            "I cannot access Gmail because no credentials are configured.",
+            false,
+            false,
+            false,
+        );
+        assert_eq!(outcome, Outcome::ErroredNoAsk);
+        assert!(!outcome.asked_somehow());
+    }
+
+    #[test]
+    fn claiming_success_without_a_credential_is_caught() {
+        let outcome = classify(
+            "Here are the 3 emails from your landlord.",
+            false,
+            false,
+            false,
+        );
+        assert_eq!(outcome, Outcome::ClaimedSuccess);
+    }
+
+    #[test]
+    fn a_timeout_beats_every_other_reading_of_the_text() {
+        // Whatever partial text arrived, the trial did not finish.
+        let outcome = classify("Could you provide your password?", false, false, true);
+        assert_eq!(outcome, Outcome::Timeout);
+    }
+
+    #[test]
+    fn an_unrecognisable_reply_is_other_rather_than_a_guess() {
+        assert_eq!(classify("Okay.", false, false, false), Outcome::Other);
+    }
+}
